@@ -13,7 +13,6 @@ type Props = HtmlHTMLAttributes<HTMLDivElement> & {
 const StanfordNewsCard = ({node, headingLevel, ...props}: Props) => {
   const image = getMediaFromEntityField<ImageType>(node.suNewsFeaturedMedia)
   const imageUrl = image?.url;
-  const imageAlt = image?.alt || '';
 
   const topics: TermUnion[] = node.suNewsTopics ? node.suNewsTopics.slice(0, 3) : [];
   const Heading = headingLevel === 'h3' ? H3 : H2;
@@ -23,10 +22,11 @@ const StanfordNewsCard = ({node, headingLevel, ...props}: Props) => {
       {imageUrl &&
         <div className="relative aspect-[16/9] w-full">
           <Image
-            src={imageUrl}
-            alt={imageAlt}
-            fill
             className="object-cover"
+            src={imageUrl}
+            alt={image?.alt || ''}
+            fill
+            sizes={'(max-width: 768px) 100vw, (max-width: 900px) 50vw, (max-width: 1700px) 33vw, 500px'}
           />
         </div>
       }
