@@ -14,7 +14,6 @@ const StanfordPageCard = ({node, headingLevel, ...props}: Props) => {
   const image = getMediaFromEntityField<ImageType>(node.suPageImage) || getMediaFromEntityField<ImageType>(node.suPageBanner?.__typename === 'ParagraphStanfordBanner' ? node.suPageBanner?.suBannerImage : undefined);
 
   const imageUrl = image?.url
-  const imageAlt = image?.alt || '';
 
   const Heading = headingLevel === 'h3' ? H3 : H2;
   return (
@@ -23,10 +22,11 @@ const StanfordPageCard = ({node, headingLevel, ...props}: Props) => {
         <div
           className="relative aspect-[16/9] w-full">
           <Image
-            src={imageUrl}
-            alt={imageAlt}
-            fill
             className="object-cover"
+            src={imageUrl}
+            alt={image?.alt || ''}
+            fill
+            sizes={'(max-width: 768px) 100vw, (max-width: 900px) 50vw, (max-width: 1700px) 33vw, 500px'}
           />
         </div>
       }
