@@ -25,11 +25,12 @@ const StanfordPersonPage = ({node, ...props}: Props) => {
         {imageUrl &&
           <div className="relative aspect-[1/1] w-[250px] shrink-0 mx-auto lg:mx-0">
             <Image
+              className="rounded-full"
               src={imageUrl}
               alt=""
-              fill
-              className="rounded-full"
               loading="eager"
+              fill
+              sizes={'(max-width: 768px) 100vw, (max-width: 900px) 50vw, (max-width: 1700px) 33vw, 500px'}
             />
           </div>
         }
