@@ -1,19 +1,13 @@
 import StanfordCourseListItem from "@components/nodes/list-item/stanford-course/stanford-course-list-item";
 import LoadMoreList from "@components/elements/load-more-list";
-import {NodeStanfordCourse, Maybe} from "@lib/gql/__generated__/drupal";
+import {NodeStanfordCourse} from "@lib/gql/__generated__/drupal";
 
 interface Props {
-  emptyMessage?: Maybe<string>
   headingLevel: string
   items?: NodeStanfordCourse[]
 }
 
-const CourseListView = async ({items = [], emptyMessage, headingLevel}: Props) => {
-
-  if (items.length === 0) {
-    return emptyMessage ? <div>{emptyMessage}</div> : null;
-  }
-
+const CourseListView = async ({items = [], headingLevel}: Props) => {
   return (
     <LoadMoreList
       buttonText={<>Load More<span className="sr-only">&nbsp;courses</span></>}
