@@ -1,16 +1,12 @@
 import CardViewGrid from "@components/views/card-view-grid";
-import {NodeStanfordPerson, Maybe} from "@lib/gql/__generated__/drupal";
+import {NodeStanfordPerson} from "@lib/gql/__generated__/drupal";
 
 interface Props {
-  emptyMessage?: Maybe<string>
   headingLevel: string
   items?: NodeStanfordPerson[]
 }
 
-const PersonCardView = async ({items = [], emptyMessage, headingLevel}: Props) => {
-  if (items.length === 0) {
-    return emptyMessage ? <div>{emptyMessage}</div> : null;
-  }
+const PersonCardView = async ({items = [], headingLevel}: Props) => {
   return (
     <CardViewGrid items={items} headingLevel={headingLevel}/>
   )
